moved some tests into test folder
